+++ docs/HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add safe_html filter for text configuration
+  [cekk]
 
 
 0.1.0 (2015-09-16)